fix(js): allow publishable library to run release command
