+++ packages/js/src/generators/library/library.ts
@@ -5,14 +5,11 @@ import {
   formatFiles,
   generateFiles,
   GeneratorCallback,
-  getPackageManagerCommand,
   installPackagesTask,
   joinPathFragments,
   names,
   offsetFromRoot,
-  output,
   ProjectConfiguration,
-  ProjectGraphProjectNode,
   readNxJson,
   readProjectConfiguration,
   runTasksInSerial,
@@ -69,6 +66,10 @@ import type {
 } from './schema';
 import { sortPackageJsonFields } from '../../utils/package-json/sort-fields';
 import { getImportPath } from '../../utils/get-import-path';
+import {
+  addReleaseOptionForPublishableTarget,
+  releaseTasks,
+} from './utils/add-release-config';
 
 const defaultOutputDirectory = 'dist';
 
@@ -113,10 +114,6 @@ export async function libraryGeneratorInternal(
 
   if (!options.skipPackageJson) {
     tasks.push(addProjectDependencies(tree, options));
-  }
-
-  if (options.publishable) {
-    tasks.push(await setupVerdaccio(tree, { ...options, skipFormat: true }));
   }
 
   if (options.bundler === 'rollup') {
@@ -249,9 +246,7 @@ export async function libraryGeneratorInternal(
   }
 
   if (options.publishable) {
-    tasks.push(() => {
-      logNxReleaseDocsInfo();
-    });
+    tasks.push(await releaseTasks(tree));
   }
 
   // Always run install to link packages.
@@ -337,32 +332,13 @@ async function configureProject(
   }
 
   if (options.publishable) {
-    if (!options.isUsingTsSolutionConfig) {
-      const packageRoot = joinPathFragments(
-        defaultOutputDirectory,
-        '{projectRoot}'
-      );
-
-      projectConfiguration.targets ??= {};
-      projectConfiguration.targets['nx-release-publish'] = {
-        options: {
-          packageRoot,
-        },
-      };
-
-      projectConfiguration.release = {
-        version: {
-          generatorOptions: {
-            packageRoot,
-            // using git tags to determine the current version is required here because
-            // the version in the package root is overridden with every build
-            currentVersionResolver: 'git-tag',
-          },
-        },
-      };
-    }
-
-    await addProjectToNxReleaseConfig(tree, options, projectConfiguration);
+    await addReleaseOptionForPublishableTarget(
+      tree,
+      options.name,
+      projectConfiguration,
+      defaultOutputDirectory,
+      options.isUsingTsSolutionConfig
+    );
   }
 
   if (!options.useProjectJson) {
@@ -1257,120 +1233,6 @@ function determineEntryFields(
       return {};
     }
   }
-}
-
-function projectsConfigMatchesProject(
-  projectsConfig: string | string[] | undefined,
-  project: ProjectGraphProjectNode
-): boolean {
-  if (!projectsConfig) {
-    return false;
-  }
-
-  if (typeof projectsConfig === 'string') {
-    projectsConfig = [projectsConfig];
-  }
-
-  const graph: Record<string, ProjectGraphProjectNode> = {
-    [project.name]: project,
-  };
-
-  const matchingProjects = findMatchingProjects(projectsConfig, graph);
-
-  return matchingProjects.includes(project.name);
-}
-
-async function addProjectToNxReleaseConfig(
-  tree: Tree,
-  options: NormalizedLibraryGeneratorOptions,
-  projectConfiguration: ProjectConfiguration
-) {
-  const nxJson = readNxJson(tree);
-
-  const addPreVersionCommand = () => {
-    const pmc = getPackageManagerCommand();
-
-    nxJson.release = {
-      ...nxJson.release,
-      version: {
-        preVersionCommand: `${pmc.dlx} nx run-many -t build`,
-        ...nxJson.release?.version,
-      },
-    };
-  };
-
-  if (!nxJson.release || (!nxJson.release.projects && !nxJson.release.groups)) {
-    // skip adding any projects configuration since the new project should be
-    // automatically included by nx release's default project detection logic
-    addPreVersionCommand();
-    writeJson(tree, 'nx.json', nxJson);
-    return;
-  }
-
-  const project: ProjectGraphProjectNode = {
-    name: options.name,
-    type: 'lib' as const,
-    data: {
-      root: projectConfiguration.root,
-      tags: projectConfiguration.tags,
-    },
-  };
-
-  if (projectsConfigMatchesProject(nxJson.release.projects, project)) {
-    output.log({
-      title: `Project already included in existing release configuration`,
-    });
-    addPreVersionCommand();
-    writeJson(tree, 'nx.json', nxJson);
-    return;
-  }
-
-  if (Array.isArray(nxJson.release.projects)) {
-    nxJson.release.projects.push(options.name);
-    addPreVersionCommand();
-    writeJson(tree, 'nx.json', nxJson);
-    output.log({
-      title: `Added project to existing release configuration`,
-    });
-  }
-
-  if (nxJson.release.groups) {
-    const allGroups = Object.entries(nxJson.release.groups);
-
-    for (const [name, group] of allGroups) {
-      if (projectsConfigMatchesProject(group.projects, project)) {
-        addPreVersionCommand();
-        writeJson(tree, 'nx.json', nxJson);
-        return `Project already included in existing release configuration for group ${name}`;
-      }
-    }
-
-    output.warn({
-      title: `Could not find a release group that includes ${options.name}`,
-      bodyLines: [
-        `Ensure that ${options.name} is included in a release group's "projects" list in nx.json so it can be published with "nx release"`,
-      ],
-    });
-    addPreVersionCommand();
-    writeJson(tree, 'nx.json', nxJson);
-    return;
-  }
-
-  if (typeof nxJson.release.projects === 'string') {
-    nxJson.release.projects = [nxJson.release.projects, options.name];
-    addPreVersionCommand();
-    writeJson(tree, 'nx.json', nxJson);
-    output.log({
-      title: `Added project to existing release configuration`,
-    });
-    return;
-  }
-}
-
-function logNxReleaseDocsInfo() {
-  output.log({
-    title: `📦 To learn how to publish this library, see https://nx.dev/core-features/manage-releases.`,
-  });
 }
 
 function findRootJestPreset(tree: Tree): string | null {
